1, use _countof instead of sizeof/sizeof 2, add string length attribute to pr file 3, modify prgen to fit for 2 4, now client can connect to gate server
vet/app: me

+++ ServerBase/server_config.h
@@ -37,7 +37,7 @@ private:
 
 public:
 	TCHAR m_strServerName[MAX_PATH+1];
-	uint16 m_iServerId;
+	uint8 m_iServerId;
 	uint32 m_iPeerIP;
 	uint16 m_iPeerPort;
